[Lua] Fix sub list serialization in JNLua and some more cleanups
